requirements.txt
ansitable==0.11.5
cfgv==3.5.0
colorama==0.4.6
colored==2.3.2
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
dynamixel-sdk==4.0.3
filelock==3.28.0
fonttools==4.62.1
identify==2.6.18
kiwisolver==1.5.0
matplotlib==3.10.8
nodeenv==1.10.0
numpy==1.26.4
packaging==26.1
pgraph-python==0.6.3
pillow==12.2.0
platformdirs==4.9.6
pre_commit==4.5.1
progress==1.6.1
pyaml==26.2.1
PyOpenGL==3.1.10
pyparsing==3.3.2
pyqtgraph==0.14.0
pyserial==3.5
PySide6==6.11.0
PySide6_Addons==6.11.0
PySide6_Essentials==6.11.0
python-dateutil==2.9.0.post0
python-discovery==1.2.2
PyYAML==6.0.3
roboticstoolbox-python==1.1.1
rtb-data==1.0.1
scipy==1.17.1
shiboken6==6.11.0
six==1.17.0
spatialgeometry==1.1.0
spatialmath-python==1.1.15
swift-sim==1.1.0
trimesh==4.11.5
typing_extensions==4.15.0
virtualenv==21.2.4
websockets==16.0
